requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==21.2.0 